+++ src/main/java/net/rptools/maptool/client/ui/campaignproperties/CampaignPropertiesDialog.java
@@ -18,23 +18,13 @@ import java.awt.*;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.awt.event.KeyEvent;
-import java.io.BufferedReader;
 import java.io.File;
 import java.io.IOException;
-import java.io.LineNumberReader;
-import java.io.StringReader;
-import java.text.ParseException;
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashMap;
-import java.util.LinkedHashMap;
-import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
-import java.util.Map.Entry;
-import java.util.TreeMap;
-import java.util.regex.Pattern;
 import javax.swing.*;
 import net.rptools.lib.FileUtil;
 import net.rptools.maptool.client.AppConstants;
@@ -53,9 +43,9 @@ import net.rptools.maptool.model.Light;
 import net.rptools.maptool.model.LightSource;
 import net.rptools.maptool.model.ShapeType;
 import net.rptools.maptool.model.SightType;
-import net.rptools.maptool.model.drawing.DrawableColorPaint;
+import net.rptools.maptool.util.LightSyntax;
 import net.rptools.maptool.util.PersistenceUtil;
-import net.rptools.maptool.util.StringUtil;
+import net.rptools.maptool.util.SightSyntax;
 
 public class CampaignPropertiesDialog extends JDialog {
   public enum Status {
@@ -241,195 +231,11 @@ public class CampaignPropertiesDialog extends JDialog {
   }
 
   private String updateSightPanel(Map<String, SightType> sightTypeMap) {
-    StringBuilder builder = new StringBuilder();
-    for (SightType sight : sightTypeMap.values()) {
-      builder.append(sight.getName()).append(": ");
-
-      builder.append(sight.getShape().name().toLowerCase()).append(" ");
-
-      switch (sight.getShape()) {
-        case SQUARE, CIRCLE, GRID, HEX:
-          break;
-        case BEAM:
-          if (sight.getArc() != 0) {
-            builder.append("arc=").append(StringUtil.formatDecimal(sight.getArc())).append(' ');
-          } else {
-            builder.append("arc=4").append(StringUtil.formatDecimal(sight.getArc())).append(' ');
-          }
-          if (sight.getOffset() != 0) {
-            builder
-                .append("offset=")
-                .append(StringUtil.formatDecimal(sight.getOffset()))
-                .append(' ');
-          }
-          break;
-        case CONE:
-          if (sight.getArc() != 0) {
-            builder.append("arc=").append(StringUtil.formatDecimal(sight.getArc())).append(' ');
-          }
-          if (sight.getOffset() != 0) {
-            builder
-                .append("offset=")
-                .append(StringUtil.formatDecimal(sight.getOffset()))
-                .append(' ');
-          }
-          break;
-      }
-      if (sight.getDistance() != 0) {
-        builder
-            .append("distance=")
-            .append(StringUtil.formatDecimal(sight.getDistance()))
-            .append(' ');
-      }
-
-      // Scale with Token
-      if (sight.isScaleWithToken()) {
-        builder.append("scale ");
-      }
-      // Multiplier
-      if (sight.getMultiplier() != 1 && sight.getMultiplier() != 0) {
-        builder.append("x").append(StringUtil.formatDecimal(sight.getMultiplier())).append(' ');
-      }
-      // Personal light
-      if (sight.getPersonalLightSource() != null) {
-        LightSource source = sight.getPersonalLightSource();
-
-        if (source.getLightList() != null) {
-          for (Light light : source.getLightList()) {
-            double range = light.getRadius();
-
-            builder.append("r").append(StringUtil.formatDecimal(range));
-
-            if (light.getPaint() != null && light.getPaint() instanceof DrawableColorPaint) {
-              Color color = (Color) light.getPaint().getPaint();
-              builder.append(toHex(color));
-            }
-            final var lumens = light.getLumens();
-            if (lumens >= 0) {
-              builder.append('+');
-            }
-            builder.append(Integer.toString(lumens, 10));
-            builder.append(' ');
-          }
-        }
-      }
-      builder.append('\n');
-    }
-    return builder.toString();
+    return new SightSyntax().stringify(sightTypeMap);
   }
 
   private String updateLightPanel(Map<String, Map<GUID, LightSource>> lightSources) {
-    StringBuilder builder = new StringBuilder();
-    for (Entry<String, Map<GUID, LightSource>> entry : lightSources.entrySet()) {
-      builder.append(entry.getKey());
-      builder.append("\n----\n");
-
-      for (LightSource lightSource : entry.getValue().values()) {
-        builder.append(lightSource.getName()).append(":");
-
-        if (lightSource.getType() != LightSource.Type.NORMAL) {
-          builder.append(' ').append(lightSource.getType().name().toLowerCase());
-        }
-        if (lightSource.isScaleWithToken()) {
-          builder.append(" scale");
-        }
-
-        final var lastParameters = new LinkedHashMap<String, Object>();
-        lastParameters.put("", null);
-        lastParameters.put("arc", 0.);
-        lastParameters.put("offset", 0.);
-        lastParameters.put("GM", false);
-        lastParameters.put("OWNER", false);
-
-        for (Light light : lightSource.getLightList()) {
-          final var parameters = new HashMap<>();
-
-          // TODO: This HAS to change, the lights need to be auto describing, this hard wiring sucks
-          if (lightSource.getType() == LightSource.Type.AURA) {
-            parameters.put("GM", light.isGM());
-            parameters.put("OWNER", light.isOwnerOnly());
-          }
-
-          parameters.put("", light.getShape().name().toLowerCase());
-          switch (light.getShape()) {
-            default:
-              throw new RuntimeException(
-                  "Unrecognized shape: " + light.getShape().toString().toLowerCase());
-            case SQUARE, GRID, CIRCLE, HEX:
-              break;
-            case BEAM:
-              parameters.put("arc", light.getArcAngle());
-              parameters.put("offset", light.getFacingOffset());
-              break;
-            case CONE:
-              parameters.put("arc", light.getArcAngle());
-              parameters.put("offset", light.getFacingOffset());
-              break;
-          }
-
-          for (final var parameterEntry : lastParameters.entrySet()) {
-            final var key = parameterEntry.getKey();
-            final var oldValue = parameterEntry.getValue();
-            final var newValue = parameters.get(key);
-
-            if (newValue != null && !newValue.equals(oldValue)) {
-              lastParameters.put(key, newValue);
-
-              // Special case: booleans are flags that are either present or not.
-              if (newValue instanceof Boolean b) {
-                if (b) {
-                  builder.append(" ").append(key);
-                }
-              } else {
-                builder.append(" ");
-                if (!"".equals(key)) {
-                  // Special case: don't include a key= for shapes.
-                  builder.append(key).append("=");
-                }
-                builder.append(
-                    switch (newValue) {
-                      case Double d -> StringUtil.formatDecimal(d);
-                      default -> newValue.toString();
-                    });
-              }
-            }
-          }
-
-          builder.append(' ').append(StringUtil.formatDecimal(light.getRadius()));
-          if (light.getPaint() instanceof DrawableColorPaint) {
-            Color color = (Color) light.getPaint().getPaint();
-            builder.append(toHex(color));
-          }
-          if (lightSource.getType() == LightSource.Type.NORMAL) {
-            final var lumens = light.getLumens();
-            if (lumens >= 0) {
-              builder.append('+');
-            }
-            builder.append(Integer.toString(lumens, 10));
-          }
-        }
-        builder.append('\n');
-      }
-      builder.append('\n');
-    }
-    return builder.toString();
-  }
-
-  private String toHex(Color color) {
-    StringBuilder builder = new StringBuilder("#");
-
-    builder.append(padLeft(Integer.toHexString(color.getRed()), '0', 2));
-    builder.append(padLeft(Integer.toHexString(color.getGreen()), '0', 2));
-    builder.append(padLeft(Integer.toHexString(color.getBlue()), '0', 2));
-
-    return builder.toString();
-  }
-
-  private String padLeft(String str, char padChar, int length) {
-    while (str.length() < length) {
-      str = padChar + str;
-    }
-    return str;
+    return new LightSyntax().stringifyCategorizedLights(lightSources);
   }
 
   private void updateRepositoryList(CampaignProperties properties) {
@@ -457,7 +263,9 @@ public class CampaignPropertiesDialog extends JDialog {
     campaign.getLightSourcesMap().clear();
     campaign.getLightSourcesMap().putAll(lightMap);
 
-    commitSightMap(getSightPanel().getText());
+    List<SightType> sightMap = commitSightMap(getSightPanel().getText());
+    campaign.setSightTypes(sightMap);
+
     tokenStatesController.copyUIToCampaign(campaign);
     tokenBarController.copyUIToCampaign(campaign);
 
@@ -470,154 +278,8 @@ public class CampaignPropertiesDialog extends JDialog {
     }
   }
 
-  private void commitSightMap(final String text) {
-    List<SightType> sightList = new LinkedList<SightType>();
-    LineNumberReader reader = new LineNumberReader(new BufferedReader(new StringReader(text)));
-    String line = null;
-    String toBeParsed = null, errmsg = null;
-    List<String> errlog = new LinkedList<String>();
-    try {
-      while ((line = reader.readLine()) != null) {
-        line = line.trim();
-
-        // Blanks
-        if (line.length() == 0 || line.indexOf(':') < 1) {
-          continue;
-        }
-        // Parse line
-        int split = line.indexOf(':');
-        String label = line.substring(0, split).trim();
-        String value = line.substring(split + 1).trim();
-
-        if (label.length() == 0) {
-          continue;
-        }
-        // Parse Details
-        double magnifier = 1;
-        // If null, no personal light has been defined.
-        List<Light> personalLightLights = null;
-
-        String[] args = value.split("\\s+");
-        ShapeType shape = ShapeType.CIRCLE;
-        boolean scaleWithToken = false;
-        int arc = 90;
-        float range = 0;
-        int offset = 0;
-        double pLightRange = 0;
-
-        for (String arg : args) {
-          assert arg.length() > 0; // The split() uses "one or more spaces", removing empty strings
-          try {
-            shape = ShapeType.valueOf(arg.toUpperCase());
-            arc = shape == ShapeType.BEAM ? 4 : arc;
-            continue;
-          } catch (IllegalArgumentException iae) {
-            // Expected when not defining a shape
-          }
-          // Scale with Token
-          if (arg.equalsIgnoreCase("SCALE")) {
-            scaleWithToken = true;
-            continue;
-          }
-          try {
-
-            if (arg.startsWith("x")) {
-              toBeParsed = arg.substring(1); // Used in the catch block, below
-              errmsg = "msg.error.mtprops.sight.multiplier"; // (ditto)
-              magnifier = StringUtil.parseDecimal(toBeParsed);
-            } else if (arg.startsWith("r")) { // XXX Why not "r=#" instead of "r#"??
-              toBeParsed = arg.substring(1);
-              errmsg = "msg.error.mtprops.sight.range";
-
-              final var rangeRegex = Pattern.compile("([^#+-]*)(#[0-9a-fA-F]+)?([+-]\\d*)?");
-              final var matcher = rangeRegex.matcher(toBeParsed);
-              if (matcher.find()) {
-                pLightRange = StringUtil.parseDecimal(matcher.group(1));
-                final var colorString = matcher.group(2);
-                final var lumensString = matcher.group(3);
-                // Note that Color.decode() _wants_ the leading "#", otherwise it might not treat
-                // the value as a hex code.
-                Color personalLightColor = null;
-                if (colorString != null) {
-                  personalLightColor = Color.decode(colorString);
-                }
-                int perRangeLumens = 100;
-                if (lumensString != null) {
-                  perRangeLumens = Integer.parseInt(lumensString, 10);
-                  if (perRangeLumens == 0) {
-                    errlog.add(
-                        I18N.getText("msg.error.mtprops.sight.zerolumens", reader.getLineNumber()));
-                    perRangeLumens = 100;
-                  }
-                }
-
-                if (personalLightLights == null) {
-                  personalLightLights = new ArrayList<>();
-                }
-                personalLightLights.add(
-                    new Light(
-                        shape,
-                        0,
-                        pLightRange,
-                        arc,
-                        personalLightColor == null
-                            ? null
-                            : new DrawableColorPaint(personalLightColor),
-                        perRangeLumens,
-                        false,
-                        false));
-              } else {
-                throw new ParseException(
-                    String.format("Unrecognized personal light syntax: %s", arg), 0);
-              }
-            } else if (arg.startsWith("arc=") && arg.length() > 4) {
-              toBeParsed = arg.substring(4);
-              errmsg = "msg.error.mtprops.sight.arc";
-              arc = StringUtil.parseInteger(toBeParsed);
-            } else if (arg.startsWith("distance=") && arg.length() > 9) {
-              toBeParsed = arg.substring(9);
-              errmsg = "msg.error.mtprops.sight.distance";
-              range = StringUtil.parseDecimal(toBeParsed).floatValue();
-            } else if (arg.startsWith("offset=") && arg.length() > 7) {
-              toBeParsed = arg.substring(7);
-              errmsg = "msg.error.mtprops.sight.offset";
-              offset = StringUtil.parseInteger(toBeParsed);
-            } else {
-              toBeParsed = arg;
-              errmsg =
-                  I18N.getText(
-                      "msg.error.mtprops.sight.unknownField", reader.getLineNumber(), toBeParsed);
-              errlog.add(errmsg);
-            }
-          } catch (ParseException e) {
-            assert errmsg != null;
-            errlog.add(I18N.getText(errmsg, reader.getLineNumber(), toBeParsed));
-          }
-        }
-
-        LightSource personalLight =
-            personalLightLights == null
-                ? null
-                : LightSource.createPersonal(scaleWithToken, personalLightLights);
-        SightType sight =
-            new SightType(label, magnifier, personalLight, shape, arc, scaleWithToken);
-        sight.setDistance(range);
-        sight.setOffset(offset);
-
-        // Store
-        sightList.add(sight);
-      }
-    } catch (IOException ioe) {
-      MapTool.showError("msg.error.mtprops.sight.ioexception", ioe);
-    }
-    if (!errlog.isEmpty()) {
-      // Show the user a list of errors so they can (attempt to) correct all of them at once
-      MapTool.showFeedback(errlog.toArray());
-      errlog.clear();
-      throw new IllegalArgumentException(
-          "msg.error.mtprops.sight.definition"); // Don't save sights...
-    }
-    campaign.setSightTypes(sightList);
+  private List<SightType> commitSightMap(final String text) {
+    return new SightSyntax().parse(text);
   }
 
   /**
@@ -645,206 +307,7 @@ public class CampaignPropertiesDialog extends JDialog {
    */
   private Map<String, Map<GUID, LightSource>> commitLightMap(
       final String text, final Map<String, Map<GUID, LightSource>> originalLightSourcesMap) {
-    Map<String, Map<GUID, LightSource>> lightMap = new TreeMap<String, Map<GUID, LightSource>>();
-    LineNumberReader reader = new LineNumberReader(new BufferedReader(new StringReader(text)));
-    String line = null;
-    List<String> errlog = new LinkedList<String>();
-
-    try {
-      String currentGroupName = null;
-      Map<GUID, LightSource> lightSourceMap = null;
-
-      while ((line = reader.readLine()) != null) {
-        line = line.trim();
-
-        // Comments
-        if (line.length() > 0 && line.charAt(0) == '-') {
-          continue;
-        }
-        // Blank lines
-        if (line.length() == 0) {
-          if (currentGroupName != null) {
-            lightMap.put(currentGroupName, lightSourceMap);
-          }
-          currentGroupName = null;
-          continue;
-        }
-        // New group
-        if (currentGroupName == null) {
-          currentGroupName = line;
-          lightSourceMap = new HashMap<GUID, LightSource>();
-          continue;
-        }
-        // Item
-        int split = line.indexOf(':');
-        if (split < 1) {
-          continue;
-        }
-
-        // region Light source properties.
-        String name = line.substring(0, split).trim();
-        GUID id = new GUID();
-        LightSource.Type type = LightSource.Type.NORMAL;
-        boolean scaleWithToken = false;
-        List<Light> lights = new ArrayList<>();
-        // endregion
-        // region Individual light properties
-        ShapeType shape = ShapeType.CIRCLE; // TODO: Make a preference for default shape
-        double arc = 0;
-        double offset = 0;
-        boolean gmOnly = false;
-        boolean owner = false;
-        String distance = null;
-        // endregion
-
-        for (String arg : line.substring(split + 1).split("\\s+")) {
-          arg = arg.trim();
-          if (arg.length() == 0) {
-            continue;
-          }
-          if (arg.equalsIgnoreCase("GM")) {
-            gmOnly = true;
-            owner = false;
-            continue;
-          }
-          if (arg.equalsIgnoreCase("OWNER")) {
-            gmOnly = false;
-            owner = true;
-            continue;
-          }
-          // Scale with token designation
-          if (arg.equalsIgnoreCase("SCALE")) {
-            scaleWithToken = true;
-            continue;
-          }
-          // Shape designation ?
-          try {
-            shape = ShapeType.valueOf(arg.toUpperCase());
-            arc = shape == ShapeType.BEAM ? 4 : arc;
-            continue;
-          } catch (IllegalArgumentException iae) {
-            // Expected when not defining a shape
-          }
-
-          // Type designation ?
-          try {
-            type = LightSource.Type.valueOf(arg.toUpperCase());
-            continue;
-          } catch (IllegalArgumentException iae) {
-            // Expected when not defining a shape
-          }
-
-          // Facing offset designation
-          if (arg.toUpperCase().startsWith("OFFSET=")) {
-            try {
-              offset = Integer.parseInt(arg.substring(7));
-              continue;
-            } catch (NullPointerException noe) {
-              errlog.add(
-                  I18N.getText("msg.error.mtprops.light.offset", reader.getLineNumber(), arg));
-            }
-          }
-
-          // Parameters
-          split = arg.indexOf('=');
-          if (split > 0) {
-            String key = arg.substring(0, split);
-            String value = arg.substring(split + 1);
-
-            // TODO: Make this a generic map to pass instead of 'arc'
-            if ("arc".equalsIgnoreCase(key)) {
-              try {
-                arc = StringUtil.parseDecimal(value);
-                shape =
-                    (shape != ShapeType.CONE && shape != ShapeType.BEAM)
-                        ? ShapeType.CONE
-                        : shape; // If the user specifies an arc, force the shape to CONE
-              } catch (ParseException pe) {
-                errlog.add(
-                    I18N.getText("msg.error.mtprops.light.arc", reader.getLineNumber(), value));
-              }
-            }
-            continue;
-          }
-
-          Color color = null;
-          int perRangeLumens = 100;
-          distance = arg;
-
-          final var rangeRegex = Pattern.compile("([^#+-]*)(#[0-9a-fA-F]+)?([+-]\\d*)?");
-          final var matcher = rangeRegex.matcher(arg);
-          if (matcher.find()) {
-            distance = matcher.group(1);
-            final var colorString = matcher.group(2);
-            final var lumensString = matcher.group(3);
-            // Note that Color.decode() _wants_ the leading "#", otherwise it might not treat the
-            // value as a hex code.
-            if (colorString != null) {
-              color = Color.decode(colorString);
-            }
-            if (lumensString != null) {
-              perRangeLumens = Integer.parseInt(lumensString, 10);
-              if (perRangeLumens == 0) {
-                errlog.add(
-                    I18N.getText("msg.error.mtprops.light.zerolumens", reader.getLineNumber()));
-                perRangeLumens = 100;
-              }
-            }
-          }
-
-          boolean isAura = type == LightSource.Type.AURA;
-          if (!isAura && (gmOnly || owner)) {
-            errlog.add(I18N.getText("msg.error.mtprops.light.gmOrOwner", reader.getLineNumber()));
-            gmOnly = false;
-            owner = false;
-          }
-          owner = gmOnly ? false : owner;
-          try {
-            Light t =
-                new Light(
-                    shape,
-                    offset,
-                    StringUtil.parseDecimal(distance),
-                    arc,
-                    color == null ? null : new DrawableColorPaint(color),
-                    perRangeLumens,
-                    gmOnly,
-                    owner);
-            lights.add(t);
-          } catch (ParseException pe) {
-            errlog.add(
-                I18N.getText("msg.error.mtprops.light.distance", reader.getLineNumber(), distance));
-          }
-        }
-        // Keep ID the same if modifying existing light. This avoids tokens losing their lights when
-        // the light definition is modified.
-        if (originalLightSourcesMap.containsKey(currentGroupName)) {
-          for (LightSource ls : originalLightSourcesMap.get(currentGroupName).values()) {
-            if (ls.getName().equalsIgnoreCase(name)) {
-              assert ls.getId() != null;
-              id = ls.getId();
-              break;
-            }
-          }
-        }
-
-        final var source = LightSource.createRegular(name, id, type, scaleWithToken, lights);
-        lightSourceMap.put(source.getId(), source);
-      }
-      // Last group
-      if (currentGroupName != null) {
-        lightMap.put(currentGroupName, lightSourceMap);
-      }
-    } catch (IOException ioe) {
-      MapTool.showError("msg.error.mtprops.light.ioexception", ioe);
-    }
-    if (!errlog.isEmpty()) {
-      MapTool.showFeedback(errlog.toArray());
-      errlog.clear();
-      throw new IllegalArgumentException(
-          "msg.error.mtprops.light.definition"); // Don't save lights...
-    }
-    return lightMap;
+    return new LightSyntax().parseCategorizedLights(text, originalLightSourcesMap);
   }
 
   public JEditorPane getLightPanel() {
